feat(duel)!: also check for opponent's status when using `/duel submit`
This commit adds a check for the opponent's status when using the
`/duel submit` command. Now, the player that invokes the command can
lose the duel if the opponent have completed solving the problem before
the player.

The same checks are also done when duel is timed out.

This commit also includes:

* Refactored commonly used operations to separate functions
  (`__get_players` and `__get_opponent`)
* Make follow-up messages/announcements more informative