feat(duel): add optional `difficulty` param and better problem selection
The optional difficulty parameter allows duel players (or assigner) to
determine the difficulty of the random problem selected for the duel.

The problem pool for duel has also been filtered with a list of allowed
topics only and a topics blacklist to prevent certain topics from being
selected.
To achieve this, another database API has been added to api.py to
retrieve a list of problems that are also attached a list of topics they
belong to.
The reason another API is written, instead of modifying the existing
one, is mostly due to the noticeably poorer performance of the new API,
probably due to the additional topics attaching operations.